requirements.txt
PyYAML==3.11
requests==2.5.3 [security]
docopt==0.6.2
loadimpact==1.1.0
tabulate==0.7.5
pygal=1.7.0
numpy=1.9.2